+++ datacontract/templates/style/tailwind.config.js
@@ -1,5 +1,5 @@
 module.exports = {
-  content: ["../datacontract.html"],
+  content: ["../datacontract.html", "../index.html"],
   theme: { },
   plugins: [],
 }