+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added test support for `azure` (#146)
 - Added support for `delta` tables on S3 (#24)
-
+- Added new command `datacontract catalog` that generates a data contract catalog with an `index.html` file.
 
 ## [0.10.1] - 2024-04-19